Creation of proxy cli sample.
Creates a own sample for testing with the
proxy cli. More or less an empty sample with
cfg_srv and health srv. prj.cfg is configured to
add the proxy cli and enable certain logging,
aswell as introducing the extended Advertiser.
Also added some extra logging to the light
sample. Finally the zephyr version points to
the corresponding commit that contains the
basic src code for the proxy cli.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>