Update of configuration model.
Adds major structural change to the cfg
model. Due to issues running the setting
set cb when using vendor model, it has
been altered to use SIG layout for the
model for now. This has no practical
impact on the implementation as far as
I can see. Further, the model has been
updated to handle and store connections
autonomusly. Connection entries can be
sent by the client to the server, and stored
in persistent memory. By using the callbacks
provided by the proxy client, the cfg server
can now reestablish connections upon
reboot or loss of connection. This feature
must be refined somewhat more, but are
working ok atm. There are some issues
regarding making proxy servers start to
advertise with node ID. I'm unsure if this
is something I have introduced to the
system, or if it is a latent bug. I will
Investigate this further.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>